+++ addons/services/tools/insert-real-items-to-s3-tricky-ids.ts
@@ -4,7 +4,7 @@
  */
 import 'dotenv/config'
 import { TxRecord } from 'shepherd-plugin-interfaces/types'
-import { processRecord } from '../services/indexer-next/src/index-shared/fnIngress/downloadWithChecks';
+import { processRecord } from '../services/indexer-next/src/index-shared/ingress/downloadWithChecks';
 import { s3HeadObject } from '../libs/utils/s3-services';
 import { destroyGatewayAgent } from '../libs/chunkStreams/gatewayStream';
 import { clearTimerHttpApiNodes } from '../libs/utils/update-range-nodes';
