+++ addons/services/tools/manually-flag-txid.ts
@@ -16,7 +16,7 @@ import { S3Client, PutObjectCommand } from '@aws-sdk/client-s3'
 import { SQSClient, SendMessageCommand, GetQueueUrlCommand } from '@aws-sdk/client-sqs'
 import { arGql } from 'ar-gql'
 import type { GQLEdgeInterface } from 'ar-gql/dist/faces'
-import { buildRecords } from '../services/indexer-next/src/index-shared/fnIngress/index'
+import { buildRecords } from '../services/indexer-next/src/index-shared/ingress/index'
 import { finalQueueName } from '../../../Config'
 import type { Config } from '../../../Config'
 import type { FilterResult } from 'shepherd-plugin-interfaces'
